requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
python-multipart==0.0.20
pydantic==2.12.5
supabase==2.28.0
google-genai==1.47.0
google-generativeai==0.8.3
gspread==6.2.1
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-api-python-client==2.191.0
pypdf==6.7.3
PyJWT==2.11.0
httpx==0.28.1
python-dotenv==1.0.0
aiohttp==3.13.3
cryptography==46.0.5
pycryptodome==3.21.0
iztro-py==0.3.4
